Refactor package to typescript
