+++ source/index.ts
@@ -1,73 +1,49 @@
-'use strict';
-const EventEmitter = require('eventemitter3');
+import EventEmitter from 'eventemitter3';
+import {Queue} from './queue';
+import PriorityQueue from './priority-queue';
+import {QueueAddOptions, DefaultAddOptions, Options} from './options';
 
-// Port of lower_bound from http://en.cppreference.com/w/cpp/algorithm/lower_bound
-// Used to compute insertion index to keep queue sorted after insertion
-function lowerBound(array, value, comp) {
-	let first = 0;
-	let count = array.length;
+type ResolveFunction<T = void> = (value?: T | PromiseLike<T>) => void;
 
-	while (count > 0) {
-		const step = (count / 2) | 0;
-		let it = first + step;
+type Task<TaskResultType> =
+		| (() => PromiseLike<TaskResultType>)
+		| (() => TaskResultType);
 
-		if (comp(array[it], value) <= 0) {
-			first = ++it;
-			count -= step + 1;
-		} else {
-			count = step;
-		}
-	}
+/**
+Promise queue with concurrency control.
+*/
+export default class PQueue<Q extends Queue<EnqueueOptionsType>, EnqueueOptionsType extends QueueAddOptions = DefaultAddOptions> extends EventEmitter {
+	private readonly _carryoverConcurrencyCount: boolean;
+	private readonly _isIntervalIgnored: boolean;
+	private _intervalCount: number;
+	private readonly _intervalCap: number;
+	private readonly _interval: number;
+	private _intervalEnd: number;
+	private _intervalId: NodeJS.Timeout | null;
+	private _timeoutId: NodeJS.Timeout | null;
 
-	return first;
-}
+	private queue: Q;
+	private readonly _queueClass: new () => Q;
 
-class PriorityQueue {
-	constructor() {
-		this._queue = [];
-	}
+	private _pendingCount: number;
+	private readonly _concurrency: number;
+	private _isPaused: boolean;
 
-	enqueue(run, options) {
-		options = {
-			priority: 0,
-			...options
-		};
+	private _resolveEmpty: ResolveFunction;
+	private _resolveIdle: ResolveFunction;
 
-		const element = {
-			priority: options.priority,
-			run
-		};
-
-		if (this.size && this._queue[this.size - 1].priority >= options.priority) {
-			this._queue.push(element);
-			return;
-		}
-
-		const index = lowerBound(this._queue, element, (a, b) => b.priority - a.priority);
-		this._queue.splice(index, 0, element);
-	}
-
-	dequeue() {
-		return this._queue.shift().run;
-	}
-
-	get size() {
-		return this._queue.length;
-	}
-}
-
-module.exports = class PQueue extends EventEmitter {
-	constructor(options) {
+	constructor(opt?: Options<Q, EnqueueOptionsType>) {
 		super();
 
-		options = {
+		const options: Options<Q, EnqueueOptionsType> = {
 			carryoverConcurrencyCount: false,
 			intervalCap: Infinity,
 			interval: 0,
 			concurrency: Infinity,
 			autoStart: true,
+			// @ts-ignore This default is a little nasty!
 			queueClass: PriorityQueue,
-			...options
+			...opt
 		};
 
 		if (!(typeof options.concurrency === 'number' && options.concurrency >= 1)) {
@@ -78,11 +54,11 @@ module.exports = class PQueue extends EventEmitter {
 			throw new TypeError(`Expected \`intervalCap\` to be a number from 1 and up, got \`${options.intervalCap}\` (${typeof options.intervalCap})`);
 		}
 
-		if (!(Number.isFinite(options.interval) && options.interval >= 0)) {
+		if (!options.interval || !(Number.isFinite(options.interval) && options.interval >= 0)) {
 			throw new TypeError(`Expected \`interval\` to be a finite number >= 0, got \`${options.interval}\` (${typeof options.interval})`);
 		}
 
-		this._carryoverConcurrencyCount = options.carryoverConcurrencyCount;
+		this._carryoverConcurrencyCount = options.carryoverConcurrencyCount!;
 		this._isIntervalIgnored = options.intervalCap === Infinity || options.interval === 0;
 		this._intervalCount = 0;
 		this._intervalCap = options.intervalCap;
@@ -91,8 +67,8 @@ module.exports = class PQueue extends EventEmitter {
 		this._intervalEnd = 0;
 		this._timeoutId = null;
 
-		this.queue = new options.queueClass(); // eslint-disable-line new-cap
-		this._queueClass = options.queueClass;
+		this.queue = new options.queueClass!(); // tslint-disable-line new-cap
+		this._queueClass = options.queueClass!;
 		this._pendingCount = 0;
 		this._concurrency = options.concurrency;
 		this._isPaused = options.autoStart === false;
@@ -144,7 +120,7 @@ module.exports = class PQueue extends EventEmitter {
 				if (this._timeoutId === null) {
 					this._timeoutId = setTimeout(() => {
 						this._onResumeInterval();
-					}, delay);
+					},                           delay);
 				}
 
 				return true;
@@ -158,7 +134,7 @@ module.exports = class PQueue extends EventEmitter {
 		if (this.queue.size === 0) {
 			// We can clear the interval ("pause")
 			// because we can redo it later ("resume")
-			clearInterval(this._intervalId);
+			clearInterval(this._intervalId!);
 			this._intervalId = null;
 
 			this._resolvePromises();
@@ -170,7 +146,7 @@ module.exports = class PQueue extends EventEmitter {
 			const canInitializeInterval = !this._intervalPaused();
 			if (this._doesIntervalAllowAnother && this._doesConcurrentAllowAnother) {
 				this.emit('active');
-				this.queue.dequeue()();
+				this.queue.dequeue()!();
 				if (canInitializeInterval) {
 					this._initializeIntervalIfNeeded();
 				}
@@ -193,16 +169,19 @@ module.exports = class PQueue extends EventEmitter {
 
 	_onInterval() {
 		if (this._intervalCount === 0 && this._pendingCount === 0) {
-			clearInterval(this._intervalId);
+			clearInterval(this._intervalId!);
 			this._intervalId = null;
 		}
 
 		this._intervalCount = (this._carryoverConcurrencyCount) ? this._pendingCount : 0;
-		while (this._tryToStartAnother()) {} // eslint-disable-line no-empty
+		while (this._tryToStartAnother()) {} // tslint-disable-line no-empty
 	}
 
-	async add(fn, options) {
-		return new Promise((resolve, reject) => {
+	/**
+	Adds a sync or async task to the queue. Always returns a promise.
+	*/
+	async add<TaskResultType>(fn: Task<TaskResultType>, options?: EnqueueOptionsType) {
+		return new Promise<TaskResultType>((resolve, reject) => {
 			const run = async () => {
 				this._pendingCount++;
 				this._intervalCount++;
@@ -221,34 +200,52 @@ module.exports = class PQueue extends EventEmitter {
 		});
 	}
 
-	async addAll(fns, options) {
+	/**
+	Same as `.add()`, but accepts an array of sync or async functions.
+	@returns A promise that resolves when all functions are resolved.
+	*/
+	async addAll<TaskResultsType>(fns: Task<TaskResultsType>[], options?: EnqueueOptionsType): Promise<TaskResultsType[]> {
 		return Promise.all(fns.map(fn => this.add(fn, options)));
 	}
 
+	/**
+	Start (or resume) executing enqueued tasks within concurrency limit. No need to call this if queue is not paused (via `options.autoStart = false` or by `.pause()` method.)
+	*/
 	start() {
 		if (!this._isPaused) {
 			return;
 		}
 
 		this._isPaused = false;
-		while (this._tryToStartAnother()) {} // eslint-disable-line no-empty
+		while (this._tryToStartAnother()) {} // tslint-disable-line no-empty
 	}
 
+	/**
+	Put queue execution on hold.
+	*/
 	pause() {
 		this._isPaused = true;
 	}
 
+	/**
+	Clear the queue.
+	*/
 	clear() {
 		this.queue = new this._queueClass();
 	}
 
+	/**
+	Can be called multiple times. Useful if you for example add additional items at a later time.
+
+	@returns A promise that settles when the queue becomes empty.
+	*/
 	async onEmpty() {
 		// Instantly resolve if the queue is empty
 		if (this.queue.size === 0) {
 			return;
 		}
 
-		return new Promise(resolve => {
+		return new Promise<void>(resolve => {
 			const existingResolve = this._resolveEmpty;
 			this._resolveEmpty = () => {
 				existingResolve();
@@ -257,13 +254,18 @@ module.exports = class PQueue extends EventEmitter {
 		});
 	}
 
+	/**
+	The difference with `.onEmpty` is that `.onIdle` guarantees that all work from the queue has finished. `.onEmpty` merely signals that the queue is empty, but it could mean that some promises haven't completed yet.
+
+	@returns A promise that settles when the queue becomes empty, and all promises have completed; `queue.size === 0 && queue.pending === 0`.
+	*/
 	async onIdle() {
 		// Instantly resolve if none pending and if nothing else is queued
 		if (this._pendingCount === 0 && this.queue.size === 0) {
 			return;
 		}
 
-		return new Promise(resolve => {
+		return new Promise<void>(resolve => {
 			const existingResolve = this._resolveIdle;
 			this._resolveIdle = () => {
 				existingResolve();
@@ -272,15 +274,24 @@ module.exports = class PQueue extends EventEmitter {
 		});
 	}
 
+	/**
+	Size of the queue.
+	*/
 	get size() {
 		return this.queue.size;
 	}
 
+	/**
+	Number of pending promises.
+	*/
 	get pending() {
 		return this._pendingCount;
 	}
 
+	/**
+	Whether the queue is currently paused.
+	*/
 	get isPaused() {
 		return this._isPaused;
 	}
-};
+}
